requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
scipy>=1.11.0
python-dotenv>=1.0.0
requests>=2.31.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytest==8.0.0
black==24.1.1
flake8==7.0.0
scikit-learn>=1.4.0
boto3>=1.34.0
botocore>=1.34.0
gunicorn>=21.2.0
psutil>=5.9.0
statsmodels>=0.14.1
numba>=0.57.0
joblib>=1.3.0
tqdm>=4.65.0 